use CEL to validate roles
While discussing the needs of exoscale/doc#1071, it became apparent
that the number of different predicates we want to handle to verify if
we are ready to allow an instance to log-in would quickly get out of
hand and force too many noisy changes in the plugin.

This patch loads a number of relevant facts about the calling instance
in a context map which can then be used to run a
[CEL](https://opensource.google/projects/cel) expression.

The following example expression are thus perfectly valid:

- `client_ip == public_ip` checks that the instance presented itself
- to the plugin with its assigned public IP `created > now -
- duration("10m")` checks that the instance was created within the
- last 10 minutes `tag["foo"] == "bar"` should be self-explanatory

With CEL, predicates can be combined using Algol/C like operators:

    client_ip == public_ip && !(tags.has("disabled")) &&
    zone == "ch-gva-2" ```

The following factoids are available in the CEL context:

- `client_ip`: The IP the request was emitted with a string
- `created`: The creation date of the instance
- `id`: The id of the instance as a string
- `name`: The name of the instance as a string
- `now`: The current timestamp
- `manager`: The manager type of this instance, if any
- `manager_id`: The ID of the manager, if any
- `public_ip`: The instance's public IP as a string
- `security_group_names`: A list of security group names the
   instance belongs to
- `security_group_ids`: A list of security group IDs the instance
   belongs to
- `tags`: A map of string to string containing the instance's tags
- `zone`: The instance's zone name
- `zone_id`: The instance's zone ID as a string

The CEL expression must be given in the `validator` field of the
`auth/exoscale` plugin's config

    $ vault write auth/exoscale/role/ci-worker \
      token_policies=ci-worker \
      validator='client_ip == public_ip && created > now - duration("10m")'

If a validator is not supplied, this expression will be stored for
evaluation:

    client_ip == public_ip

A conscious choice has been made to keep the use of the role name and
not its ID, while it could help with guessing the name of other app
roles, the checks that CEL will enable can help constrain a machine to
its role and thus avoid bearing the cost of the UX impediment it would
bring with it.