Add support for AppRole auth method compatibility
This change adds support for an AppRole-compatible authentication mode,
where Vault clients can provide the same parameters as the one required
by the `AppRole` authentication method (`role_id`/`secret_id`) in order
to work as a drop-in replacement when used in conjunction to the Vault
Agent's "Auth-Auth" functionality.

As a result of this change, the Exoscale zone must now be set at
backend-level instead of passed by the client when authenticating.